move helper functions into diffEditorWidget.ts
